requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.1
django-autoslug==1.9.9
django-cors-headers==4.4.0
django-countries==7.6.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0
